requirements.txt
pyusb==1.2.1
pyserial==3.5
opencv-python==4.10.0.84
nidaqmx==1.0.1
matplotlib==3.9.2
scipy==1.15.1